+++ nexusvault/src/main/java/nexusvault/archive/ArchiveException.java
@@ -1,0 +1,23 @@
+package nexusvault.archive;
+
+public class ArchiveException extends RuntimeException {
+
+	private static final long serialVersionUID = 3012161402564991003L;
+
+	public ArchiveException() {
+		super();
+	}
+
+	public ArchiveException(String s) {
+		super(s);
+	}
+
+	public ArchiveException(String message, Throwable cause) {
+		super(message, cause);
+	}
+
+	public ArchiveException(Throwable cause) {
+		super(cause);
+	}
+
+}