+++ nexusvault/src/main/java/nexusvault/archive/IdxEntry.java
@@ -1,0 +1,19 @@
+package nexusvault.archive;
+
+public interface IdxEntry {
+
+	String getName();
+
+	IdxDirectory getParent();
+
+	String fullName();
+
+	boolean isFile();
+
+	boolean isDir();
+
+	IdxFileLink asFile();
+
+	IdxDirectory asDirectory();
+
+}