+++ nexusvault/src/main/java/nexusvault/archive/IdxEntryNotADirectory.java
@@ -1,0 +1,23 @@
+package nexusvault.archive;
+
+public final class IdxEntryNotADirectory extends IdxException {
+
+	private static final long serialVersionUID = 5530467723975048345L;
+
+	public IdxEntryNotADirectory() {
+		super();
+	}
+
+	public IdxEntryNotADirectory(String s) {
+		super(s);
+	}
+
+	public IdxEntryNotADirectory(String message, Throwable cause) {
+		super(message, cause);
+	}
+
+	public IdxEntryNotADirectory(Throwable cause) {
+		super(cause);
+	}
+
+}