+++ nexusvault/src/main/java/nexusvault/archive/IdxFileLink.java
@@ -1,0 +1,17 @@
+package nexusvault.archive;
+
+public interface IdxFileLink extends IdxEntry {
+
+	byte[] getShaHash();
+
+	int getFlags();
+
+	long getUncompressedSize();
+
+	long getCompressedSize();
+
+	String getFileEnding();
+
+	String getNameWithoutFileEnding();
+
+}