+++ nexusvault/src/main/java/nexusvault/archive/impl/VaultUnpacker.java
@@ -1,0 +1,10 @@
+package nexusvault.archive.impl;
+
+import java.nio.ByteBuffer;
+
+import kreed.io.util.BinaryReader;
+import nexusvault.archive.VaultUnpackException;
+
+interface VaultUnpacker {
+	ByteBuffer unpack(BinaryReader reader, long compressedSize, long uncompressedSize) throws VaultUnpackException;
+}