+++ nexusvault/src/main/java/nexusvault/archive/util/IdxCollector.java
@@ -1,0 +1,9 @@
+package nexusvault.archive.util;
+
+import java.util.List;
+
+import nexusvault.archive.IdxEntry;
+
+public interface IdxCollector<T extends IdxEntry> {
+	List<T> getAndClearResult();
+}