+++ nexusvault/src/main/java/nexusvault/pack/VaultUnpacker.java
@@ -1,9 +1,0 @@
-package nexusvault.pack;
-
-import java.nio.ByteBuffer;
-
-import kreed.io.util.BinaryReader;
-
-interface VaultUnpacker {
-	ByteBuffer unpack(BinaryReader reader, long compressedSize, long uncompressedSize) throws VaultUnpackException;
-}