+++ nexusvault/src/main/java/nexusvault/pack/util/IdxCollector.java
@@ -1,9 +1,0 @@
-package nexusvault.pack.util;
-
-import java.util.List;
-
-import nexusvault.pack.index.IdxEntry;
-
-public interface IdxCollector<T extends IdxEntry> {
-	List<T> getAndClearResult();
-}