+++ nexusvault/docs/CHANGELOG.md
@@ -2,7 +2,14 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com)
 
-## [Unreleased]
+## [Unreleased] 
+
+## [2.0.0]
+### Changed
+- Package naming: nexusvault.pack -> nexusvault.archive
+- Moved public classes & interfaces to nexusvault.archive
+- Moved implementation to nexusvault.impl
+- Moved .index & .archive file structs to nexusvault.struct
 
 ## [1.0.0] - 2019-01-28
 ### Added